カスタムフックを作成 #87
